Move Bard enabled check to BardService
